clk test v4

+++ fastComm.C
@@ -39,7 +39,7 @@ void FastComm::write_tfc(uint32_t length, uint32_t command[], int period, bool s
 
 void FastComm::read_daq(uint32_t clock_delay, uint32_t length, int trigger, uint8_t *packet[1024*5]) {
 
-  uint32_t data =0;
+  uint8_t data =0;
   uint32_t e0 = 0;
   uint32_t e1 = 0;
   uint32_t e2 = 0;
